[Common][PowerDisplay][QuickAccess] Shared flyout positioning helper
Introduces FlyoutWindowHelper in Common.UI.Controls and migrates both
PowerDisplay and QuickAccess to it. Fixes two pre-existing flyout bugs:

1. PowerDisplay flyout could overlap the taskbar at 100% scaling, because
   the previous PowerDisplay-only positioning math anchored to the screen
   bounds rather than the work area on certain monitor configurations.

2. QuickAccess flyout rendered with the wrong physical size (and ended up
   partially off-screen) after switching system scaling between 150% and
   100%. The previous code passed WindowEx.Width/Height into MoveAndResize
   on every summon; those properties are computed live from the current
   physical size and current DPI, so they drift across DPI transitions
   and were then double-scaled by MoveAndResize. The window now caches the
   XAML design size once at construction and uses that as the source of
   truth.

The shared helper:
- Uses absolute screen coordinates against DisplayArea.WorkArea so it
  handles non-primary and negatively-positioned monitors correctly.
- Performs a 1x1 'teleport' onto the target display before the final
  MoveAndResize call. This avoids crossing a DPI boundary during the
  visible-size call, so WM_DPICHANGED never fires on the rendered window
  and the previous WndProc-subclass workaround is no longer required.
- Exposes overloads for bottom-right anchoring (used by both flyouts)
  and centered placement (used by PowerDisplay's IdentifyWindow).

Cleanup:
- Removes the PowerDisplay-only DpiSuppressor (the DPI-suppression code
  path is now dead). The WndProc-subclass piece that routed WM_HOTKEY
  into HotkeyService is preserved as a generic WindowMessageHook in
  Common.UI.Controls/Window/, since PowerDisplay still needs in-process
  hotkey handling.
